requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiosqlite==0.20.0
python-dotenv==1.0.1
redis==5.0.7
SQLAlchemy==2.0.31
googletrans==3.1.0a0
SunoAI==1.0.7
